push EF

+++ EnkaDotNet/Exceptions/PlayerNotFoundException.cs
@@ -4,19 +4,19 @@ namespace EnkaDotNet.Exceptions
 {
     public class PlayerNotFoundException : EnkaNetworkException
     {
-        public int Uid { get; }
+        public long Uid { get; }
 
-        public PlayerNotFoundException(int uid) : base($"Player profile for UID {uid} not found.")
+        public PlayerNotFoundException(long uid) : base($"Player profile for UID {uid} not found.")
         {
             Uid = uid;
         }
 
-        public PlayerNotFoundException(int uid, string message) : base(message)
+        public PlayerNotFoundException(long uid, string message) : base(message)
         {
             Uid = uid;
         }
 
-        public PlayerNotFoundException(int uid, string message, Exception innerException) : base(message, innerException)
+        public PlayerNotFoundException(long uid, string message, Exception innerException) : base(message, innerException)
         {
             Uid = uid;
         }
