+++ EnkaDotNet/Exceptions/ProfilePrivateException.cs
@@ -4,19 +4,19 @@ namespace EnkaDotNet.Exceptions
 {
     public class ProfilePrivateException : EnkaNetworkException
     {
-        public int Uid { get; }
+        public long Uid { get; }
 
-        public ProfilePrivateException(int uid) : base($"Profile for UID {uid} is private or character details are hidden.")
+        public ProfilePrivateException(long uid) : base($"Profile for UID {uid} is private or character details are hidden.")
         {
             Uid = uid;
         }
 
-        public ProfilePrivateException(int uid, string message) : base(message)
+        public ProfilePrivateException(long uid, string message) : base(message)
         {
             Uid = uid;
         }
 
-        public ProfilePrivateException(int uid, string message, Exception innerException) : base(message, innerException)
+        public ProfilePrivateException(long uid, string message, Exception innerException) : base(message, innerException)
         {
             Uid = uid;
         }
